requirements.txt
py-cord
discord.py
discord-together